refactoring: moves util/labels.go to model/labels.go
