--- pkg/util/labels.go
@@ -1,35 +1,0 @@
-package util
-
-import (
-	"github.com/cespare/xxhash/v2"
-	"github.com/prometheus/prometheus/model/labels"
-)
-
-var seps = []byte{'\xff'}
-
-// StableHash is a labels hashing implementation which is guaranteed to not change over time.
-// This function should be used whenever labels hashing backward compatibility must be guaranteed.
-func StableHash(ls labels.Labels) uint64 {
-	// Use xxhash.Sum64(b) for fast path as it's faster.
-	b := make([]byte, 0, 1024)
-	for i, v := range ls {
-		if len(b)+len(v.Name)+len(v.Value)+2 >= cap(b) {
-			// If labels entry is 1KB+ do not allocate whole entry.
-			h := xxhash.New()
-			_, _ = h.Write(b)
-			for _, v := range ls[i:] {
-				_, _ = h.WriteString(v.Name)
-				_, _ = h.Write(seps)
-				_, _ = h.WriteString(v.Value)
-				_, _ = h.Write(seps)
-			}
-			return h.Sum64()
-		}
-
-		b = append(b, v.Name...)
-		b = append(b, seps[0])
-		b = append(b, v.Value...)
-		b = append(b, seps[0])
-	}
-	return xxhash.Sum64(b)
-}
